feat: new structure Homework #7 Task#0 Task#1
