+++ src/PokemonCard/style.module.css
@@ -1,10 +1,3 @@
-.root {
-    width: 268px;
-    height: 388px;
-    margin: 16px;
-    perspective: 1000px;
-}
-
 .pokemonCard {
     --fire: #FDDFDF;
     --grass: #DEFDE0;
@@ -21,17 +14,29 @@
     --fighting: #E6E0D4;
     --normal:#F5F5F5;
 
-    width: inherit;
-    height: inherit;
+    width: 100%;
+    height: 100%;
     position: relative;
     transform-style: preserve-3d;
     transition: .6s .1s;
     transform: rotateY(180deg);
 }
 
-/* TODO: This place where we should change :hover on .active */
 .pokemonCard.active {
-    transform: rotateY(0deg);
+    animation: flipFront 1000ms 300ms forwards;
+}
+
+@keyframes flipFront {
+    from {
+        transform: rotateY(180deg);
+    }
+    to {
+        transform: rotateY(0deg);
+    }
+}
+
+.pokemonCard.selected {
+    margin-top: -40px;
 }
 
 .cardFront,
@@ -83,6 +88,10 @@
 
 .wrap.back {
     z-index: -1;
+    background-image: url(./../assets/card-back-side.jpg);
+    background-size: cover;
+    background-position: center;
+    background-repeat: no-repeat;
 }
 
 .cardBack {
