+++ src/routes/GamePage/routes/StartPage/style.module.css
@@ -1,0 +1,10 @@
+.flex {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-wrap: wrap;
+}
+
+.center {
+  margin: 50px 0 50px 40%;
+}
